Magnb/env file setup

+++ provision.sh
@@ -41,6 +41,15 @@ sudo usermod -aG docker root
 sudo systemctl start docker
 sudo systemctl enable docker
 
+# EXTRA CHECK TO SEE IF THAT ENV FILE IS THERE (I PRAY)
+if [ ! -f /minitwit.env ]; then 
+    echo "Missing environment file. Copy it brother just as written in the ".env.template" file"
+    exit 1
+fi
+
+# THIS KEEPS THE FILE SECURED
+sudo chmod 600 /minitwit/.env
+
 # 4. SET UP NGINX WITH HTTP CONFIG
 cp /minitwit/nginx-http.conf /minitwit/nginx.conf
